Use message queue for medium group chat messages
