+++ ts/session/crypto/MessageEncrypter.ts
@@ -2,6 +2,7 @@ import { EncryptionType } from '../types/EncryptionType';
 import { SignalService } from '../../protobuf';
 import { UserUtil } from '../../util';
 import { CipherTextObject } from '../../../libtextsecure/libsignal-protocol';
+import { encryptWithSenderKey } from '../../session/medium_group/ratchet';
 import { PubKey } from '../types';
 
 /**
@@ -29,6 +30,11 @@ function getPaddedMessageLength(originalLength: number): number {
   return messagePartCount * 160;
 }
 
+type EncryptResult = {
+  envelopeType: SignalService.Envelope.Type;
+  cipherText: Uint8Array;
+};
+
 /**
  * Encrypt `plainTextBuffer` with given `encryptionType` for `device`.
  *
@@ -41,14 +47,11 @@ export async function encrypt(
   device: PubKey,
   plainTextBuffer: Uint8Array,
   encryptionType: EncryptionType
-): Promise<{
-  envelopeType: SignalService.Envelope.Type;
-  cipherText: Uint8Array;
-}> {
+): Promise<EncryptResult> {
   const plainText = padPlainTextBuffer(plainTextBuffer);
 
   if (encryptionType === EncryptionType.MediumGroup) {
-    throw new Error('MediumGroup should not be encypted here');
+    return encryptForMediumGroup(device, plainTextBuffer);
   }
 
   const address = new window.libsignal.SignalProtocolAddress(device.key, 1);
@@ -66,6 +69,47 @@ export async function encrypt(
   }
 
   return encryptUsingSealedSender(device, innerCipherText);
+}
+
+export async function encryptForMediumGroup(
+  device: PubKey,
+  plainTextBuffer: Uint8Array
+): Promise<EncryptResult> {
+
+  const ourKey = await UserUtil.getCurrentDevicePubKey() as string;
+
+  // "Device" does not really make sense for medium groups, but
+  // that's where the group pubkey is currently stored
+  const groupId = device.key;
+
+  const { ciphertext, keyIdx } = await encryptWithSenderKey(
+    plainTextBuffer,
+    groupId,
+    ourKey
+  );
+
+  // We should include ciphertext idx in the message
+  const content = SignalService.MediumGroupCiphertext.encode({
+    ciphertext,
+    source: ourKey,
+    keyIdx,
+  }).finish();
+
+  // Encrypt for the group's identity key to hide source and key idx:
+  const {
+    ciphertext: ciphertextOuter,
+    ephemeralKey,
+  } = await window.libloki.crypto.encryptForPubkey(groupId, content);
+
+  const contentOuter = SignalService.MediumGroupContent.encode({
+    ciphertext: ciphertextOuter,
+    ephemeralKey: new Uint8Array(ephemeralKey),
+  }).finish();
+
+  const envelopeType = SignalService.Envelope.Type.MEDIUM_GROUP_CIPHERTEXT;
+
+  return { envelopeType, cipherText: contentOuter };
+
 }
 
 async function encryptUsingSealedSender(
